refactor(#169): move wikilinks MarkdownIt plugin into own file
