+++ .eleventy.js
@@ -86,10 +86,10 @@ module.exports = function (eleventyConfig) {
   let markdownItAnchor = require("markdown-it-anchor");
   let markdownFootnote = require("markdown-it-footnote");
 
-  eleventyConfig.on('eleventy.after', async () => {
-    const all = linkMapCache.all();
-    const n = 1;
-  });
+  // eleventyConfig.on('eleventy.after', async () => {
+  //   const all = linkMapCache.all();
+  //   const n = 1;
+  // });
 
   eleventyConfig
     .setLibrary("md", markdownIt({
@@ -99,27 +99,5 @@ module.exports = function (eleventyConfig) {
     }).use(markdownItAnchor, {
       permalink: false,
       slugify: input => slugify(input)
-    }).use(markdownFootnote).use(function(md) {
-      // Recognize Mediawiki links ([[text]])
-      md.linkify.add("[[", {
-        validate: /^\s?([^\[\]\|\n\r]+)(\|[^\[\]\|\n\r]+)?\s?\]\]/,
-        normalize: match => {
-          const parts = match.raw.slice(2, -2).split("|");
-          const slug = slugify(parts[0].replace(/.(md|markdown)\s?$/i, "").trim());
-          const found = linkMapCache.get(slug);
-
-          if (!found) {
-            throw new Error(`Unable to find page linked by wikilink slug [${slug}]`)
-          }
-
-          match.text = parts.length === 2
-            ? parts[1]
-            : found.title;
-
-          match.url = found.permalink.substring(0,1) === '/'
-            ? found.permalink
-            : `/${found.permalink}`;
-        }
-      })
-    }));
+    }).use(markdownFootnote).use(require('./utils/helpers/wikilinks'), linkMapCache));
 };
